requirements.txt
rich>=13.0
anthropic>=0.40.0
python-dotenv>=1.0.0
streamlit>=1.32.0
